roles/munin: enable temp and fan sensor support
We can use lm-sensors and hddtemp to provide information about CPU
and hard drive temperatures in munin-node. We should have done this
years ago, but I didn't know it existed (and it's not enabled with
the "auto" plugin group). This only makes sense to run on physical
bare metal servers, which I check with ansible_virtualization_role.
Despite its name this includes all bare metal hosts that are not
virtual machine guests!