roles/compute: update slurm.conf
As of SLURM 19.05 there is a new resource selection algorithm called
cons_tres. According to the release notes:

> offers similar functionality to cons_res with an entirely new code
> base and far greater GPU scheduling flexibility.

Also, switch from CR_CPU to CR_Core, which treats *cores* as a cons-
umable resource. This means that multiple jobs will not be allocated
on the same core to ensure no contention with on-core cache, reduce
context switching, etc.

Both of these settings are the new defaults as of 19.05 (we are on
20.11).

See: https://slurm.schedmd.com/archive/slurm-19.05.0/news.html