group_vars/dspace: adjust PostgreSQL config
Update some PostgreSQL parameters based on advice from PGTune[0] and
postgresqltuner.pl[1], as well as the annotated postgresql.conf[2].

Most importantly, I re-configured effective_cache_size to be ~37.5%
of system RAM to leave the rest for JVM heap and filesystem cache.
From there, the shared_buffers are set to ~25% of the effective
cache size (not total system RAM).

I added a few variables with their defaults commented out so that I
can remember to check them after running the server for a few days.

[0] https://pgtune.leopard.in.ua
[1] postgresqltuner.pl
[2] https://github.com/jberkus/annotated.conf/blob/master/postgresql.10.simple.conf