Add support for fail2ban's apache-botsearch
We are already using the nginx-botsearch filter on our DSpace hosts.
This allows us to use the Apache HTTPD filter on other servers where
bots might be requesting non-existent scripts like wp-login.php. If
the script is not found then the bot gets banned. It's brilliant.

For now I won't enable it on all web hosts since I don't have a way
to guarantee that each one is using Apache HTTPD or where the logs
are located. In the future I could perhaps use Ansible to check if
the apache2 or whatever package is installed, but that also means
I need to know whether it's an Ubuntu or CentOS machine.