roles/common: fix Apache fail2ban support
It turns out that the Apache HTTPD support in fail2ban is broken now
for a few years due to a change in Apache's logging behavior. Now we
must parse the access_log for HTTP 404 responses instead of looking
for "not found" in the error_log. This requires overriding the defa-
ult apache-botsearch.conf.

I have chosen to do this only on hosts with newer fail2ban versions
above 0.10.0 or so, which currently excludes Ubuntu 16.04 (which is
no longer an LTS release anyways).

Lastly, this is still opt in by setting extra_fail2ban_filters in
host_vars because I still don't have a good way to know which web
server a host has installed... that might change in the future if
I improve the Ansible web role here.

See: https://github.com/fail2ban/fail2ban/issues/2378