roles/web: generalize Let's Encrypt pre/post scripts
Assume that all hosts in the web group are running Apache HTTPD by
default, but allow it to be overriden if hosts are running nginx.

For now this doesn't do anything other than inform the Ansible te-
mplate used for the Let's Encrypt pre/post hooks. We want to make
sure that hosts have the correct stop/start scripts.