roles/compute: don't clean symbolic links in tmpwatch
We need to be extra careful with this. In my test just now I noticed
that tmpwatch wanted to remove a link where the target's last modif-
ied date was yesterday, but the link's last modified date was over 90
days ago. Touching the link and the target didn't change the behavior
so let's skip symbolic links for now (-l).

Also, the --all flag seems excessive. tmpwatch already removes files,
symbolic links, and directories by default, so that should cover most
of our needs (mostly we are concerned about large files).