roles/common: add new Apache filter
Add a new Apache filter for strange attack type that seems to abuse
open proxies. Requests of HTTP method types like GET/POST/CONNECT
that do not start with a forward slash (aka they are not relative
to the current virtual hosts's document root).

This filter is not part of the fail2ban distribution so I've copied
it as a local one.

Based on: https://www.fail2ban.org/wiki/index.php/HOWTO_apache_proxy_filter