roles/common: update fail2ban apache-proxy filter
Re-write the failregex to match the log line more accurately, as
I noticed many IPs like 0.0.0.0 and 0.0.0.1 being banned. This is
due to the default "usedns" feature doing reverse DNS lookups for
strings in the log that it mistakenly identified as hostnames.

Now the regex uses <ADDR> instead of <HOST> since we are not exp-
ecting hostnames in this logfile. I moved the <ADDR> to the front
where it belongs, added lazy quantifiers to be more precise, and
simplified the regex so it doesn't try to match the date/time (as
fail2ban strips the date/time before checking against failregex,
so all we need to do is match the literal "[]").

Lastly, I will also disable "usedns" in all Apache jails just to be
safe.