roles/common: override fail2ban's botsearch-common.conf
The botsearch-common.conf file shipped with fail2ban versions 0.10,
0.11, and newer contain a pattern intended to match files that are
part of the phpmyadmin distribution. Unfortunately this causes some
false positives, for example the Google bot made a request for the
file called "coopman.pdf" which returned an HTTP 404 and fail2ban
banned it. The problem is that the phpmyadmin pattern includes the
string "pma", which is not very unique.

So I am overriding the botsearch-common.conf file with the "pma"
removed from the phpmyadmin pattern.