roles/common: adjust fail2ban botsearch-common.local.j2
Another false positive is "(ext)?mail" in the webmail pattern. This
optionally matches "mail". I think that's a bit too broad on a busy
server with many different sites. In this case the server mirrors a
Linux package repository and some old links to the haskell-mime-mail
package cause the requesting client to get banned.