roles/common: use pyinotify backend in fail2ban
For some reason now fail2ban is trying to use the systemd backend
on some hosts where it was previously using polling. Anyways, I
realized that we should explicitly set it to use pyinotify instead
of polling because it's much more efficient. On CentOS 7 it seems
we need to manually install python-inotify, but on Ubuntu hosts
it is already installed as a dependency everywhere I checked.

See: https://www.spinics.net/lists/fail2ban/msg00110.html