roles/metrics: increase vmagent buffer
If vmagent can't connect to the VictoriaMetrics server it will keep
this much data in its local cache and send it when the server beco-
mes available again. The default is zero, which causes gaps in the
stats.

According to a 2020 blog post VictoriaMetrics uses 0.3 bytes per
sample. The node_exporter exposes about 3,000 metrics:

    # curl -s 127.0.0.1:9100/metrics | grep -cv '^#'
    ...
    3026

vmagent scrapes them every 15 seconds, so in an hour that's about:

    3000 metrics × 4 scrapes/minute × 60 minutes = 720000

And if every metric uses about 0.3 bytes, then we'd need:

    720000 * 0.3 = 216000 bytes, aka 200KB

It seems hard to believe, but I guess we can revist this later.

See: https://valyala.medium.com/prometheus-vs-victoriametrics-benchmark-on-node-exporter-metrics-4ca29c75590f