roles/dspace: remove custom certbot-auto stuff
We were using this before Ubuntu 20.04 when the standalone certbot
script from eff.org was still the recommended way. Now I have mig-
rated all hosts to Ubuntu 20.04 so we use the certbot from Ubuntu's
repos, which comes with its own timer for renewal.

Eventually we need to move to another ACME client because certbot is
not distributed outside of the Ubuntu Snap store now.