roles/common: configure EPEL and EPEL Next on CentOS Stream
Note that the ILRI EPEL repository mirrors exist, but we are not
using them currently. We only need this file to exist so we can
ignore the slurm packages in EPEL.

Also, for now we can let this run on any CentOS host, whether or
not it is in the compute/storage groups (as that's the only two
places we use them anyways, and I'm currently doing some testing
of CentOS 8 Stream).