roles/common: add script to report IPs to AbuseIPDB.com
This filters the SSH logs for hosts attempting to log in with invalid
user names, and filters out internal hosts just in case we can some
of our users fat fingering. I use this because I don't want to set up
the Fail2Ban integration. The script still needs to be added to cron.