roles: fix a few no-handler issues
These were raised by ansible-lint. In most cases we really do need
to do something when the previous task was changed, like systemd's
daemon reload in order to pick up new units or changes to existing
units, or to forcibly update the apt cache. Handlers don't work in
these cases because they run at the end, and if we "flush" them it
will cause *all* pending handlers to run!