roles/metrics: rework handlers
The forced "flush_handlers" always bothered me because it will run
all handlers that have been notified up until this point, while we
really only wanted to reload systemd in this case in order to pick
up changes to unit files. Then there was a chicken and egg problem
where we are notifying restart/reload handlers and also trying to
start the service for the first time, which only works if systemd
has been reloaded. I realized that we don't need a separate "start
and enable" task at the end because the restart handler can just
use "enabled: true".

Lastly, I can try to clean this up slightly by using a conditional
on the reload handlers, as we only want to run them if the config
has changed, not when it is new. When it is new then the restart
handler will restart the service anyways, so there's no need to
reload it. I achieved this by registering the output of the task
and checking the "diff" to see if it has 0 or more elements.