roles/compute: set autodetect for SLURM cgroup support
Both CentOS 7 and CentOS Stream 8 have cgroup/v1 by default, but
CentOS Sream 8 can be configured with cgroup/v2 if required. As of
SLURM 22.05 cgroup/v2 is supported. We are currently on 21.08 and
the "autodetect" parameter is supported, so let's use that already.

See: https://slurm.schedmd.com/cgroups.html