roles/compute: start and enable slurmd
I was wrong in 9aed0fe61c275ef1eeea309b760b5ff6a229b93f, we need to
explicitly start and enable slurmd. The restart handler *can* do it,
but there are some situations in which the config files don't change
and we end up not starting slurmd.

I will have to see how we handle the situation of needing the head
node to have slurmdbd and slurmctld started first...