roles/common: do not use fail2ban on CentOS Stream 8
It seems that fail2ban has a hard dependency on firewalld in CentOS
Stream 8, even if it can be configured to use nftables. That is not
very nice for us so we will just not use it for now. I might try to
use sshguard instead because it supports nftables.