roles/common: enable cgroups v2 on CentOS Stream 8
Use a clever trick with Ansible's lineinfile to set the pattern to
be absent, but in check mode, so it doesn't actually remove it if
found. Then we check the registered variable to see if the line was
found or not and add it. Finally we re-generate the grub config.

See: https://stackoverflow.com/questions/30786263/only-check-whether-a-line-present-in-a-file-ansible