roles: update all apt stuff
As of Debian 12 and Ubuntu 22.04 apt repositories must be signed and
apt-key is deprecated. The recommended way to specify that repos are
signed is using the `signed-by` option set in their repo file, which
points to an ASCII de-armored key in either /usr/share/keyrings or
/etc/apt/keyrings (the latter is apparently where we should place
keys that will be managed locally).

See: https://wiki.debian.org/DebianRepository/UseThirdParty