roles/common: don't run fail2ban handlers on CentOS
Newer hosts like Ubuntu 18.04+ and CentOS Stream 8+ share nftables
and fail2ban tasks, but recently I realized that fail2ban on CentOS
Stream 8 has a dependency on firewalld, so we do not run it there.

We need to only run this handler when the host is not CentOS (for
now).