roles: update method for apt keys
ansible-lint pointed out that we should use get_url instead of curl
to download these, which presents an interesting problem: it is not
clear whether we *need* to de-armor GPG keys or not. What is clear
is that we can use the keys as is as long as they are *not* named
with ".gpg" at the end, as that seems to indicate that they are in
the binary key format. I have tested this on Ubuntu 20.04 and 22.04.

Also, according to Debian's wiki on third-party repositories, the
correct location for GPG keys that are *not* managed by packages in
the given repository is /etc/apt/keyrings, but this doesn't exist
on Ubuntu 20.04 and I'm too lazy to add tasks for that right now.
The /usr/share/keyrings directory is technically for GPG keys that
are shipped by the developer in packages in the repository, thouh
I don't see nginx, Node.js, GoAccess, or PostgreSQL doing that yet.

See: https://wiki.debian.org/DebianRepository/UseThirdParty