roles/common: create /etc/apt/keyrings
According the the Debian docs for third-party repositories we must
create this manually on distros before Debian 12 and Ubuntu 22.04.

This is due to changes in apt-secure and the deprecation of apt-key.
The intended use is for user-trusted keys to be installed here, and
for keys shipped by packages to be installed in /usr/share/keyrings.

See: https://wiki.debian.org/DebianRepository/UseThirdParty
See: https://www.reddit.com/r/ansible/comments/u2vy26/apt_key/ibtvm1e/?context=3