roles/common: more updates to backup_mysql.py.j2
This is a big update:

- Remove "ignore errors" feature since I haven't tested it and would
not want it on a production server anyways.

- Re-work subprocesses to something more manageable because the ori-
ginal version by James Oguya was clever, but relied on shell=True.
Now using subprocess.run directly instead of the more complex Popen,
and use shlex to parse the command line into args that I pass to sb.

- Switch from gzip to zstd. In my testing it uses way less CPU time
and the resulting compression is much smaller. For backing up about
150 different databases individually the difference is:
  - 877 seconds for gzip, 555 seconds for zstd
  - 3 GB for gzip, 2.8 GB for zstd

- Lastly, I had hoped to make clever use of Python subprocesses by
piping the output of mysqldump directly to zstd's stdin (input) in
subprocess.run, but I was getting UTF-8 errors while backing up one
particular database:

> UnicodeDecodeError: 'utf-8' codec can't decode byte 0xa0 in position 4041588: invalid start byte

So for now I tell zstd to write the file itself.