roles/dspace: add nginx proxy cache for rest requests
Try to lighten the load on the the server caused by the REST API.
This attempts to cache REST API requests that are made by known
bots/harvesters, for which we are not interested in logging stats
hits. By my count, some REST API URLs are requested by different
bots four, five, and even six times throughout the day so we can
cache those responses.

I tested this nginx location regex against a list of URLs that we
want to match on regex101.com.

See: https://regex101.com/r/vPz11y/1