roles/dspace: implement cache bypass for some requests
If a host has sent a JSESSIONID cookie with their REST API request
then it is likely that they have already logged in and are actually
interested in getting fresh data from the REST API, as opposed to
just crawling or doing a daily harvest. In this case we can bypass
the nginx proxy cache.