misc_plays/upgrade-slurm.yml: fix updates
There is a known behavior of Ansible's yum/dnf modules that causes
"update_only: true" to work unexpectadely when the package name is
an expression like "slurm-*" that matches more packages than we h-
ave installed. The point would be to update all packages that are
installed, but it sees this as causing non-update installs, and b-
ails out. The solution here is to explicitly name the packages we
want to update. Sigh...