roles/common: update regex in parse_logs.py.j2
We want to match more failed login attempts. For example, if the
system does not have a user "ftpuser" then sshd will log:

    Invalid user ftpuser from 178.60.204.50 port 50708

Whereas, if the system has a user but it is not allowed to log in
via SSH then sshd will log:

    Failed password for invalid user root from 61.177.173.31 port 22182 ssh2

The regular expression was formerly only capturing the first one.