roles/common: adjust fail2ban filter logic
I'm getting more false positives than I'm comfortable with the
apache-proxy filter on certain web servers. I reworked the logic
used to enable extra fail2ban filters by letting it be a list,
where you can enable apache-botsearch and or apache-proxy.

To accomodate this small change I have to flatten the list to a
string before checking if apache is "in" it.