roles/dspace: copy Solr configsets
In order for the Solr user to be able to copy these, it needs to
be able to traverse the DSpace user's home directory. In Ansible
we must use the synchronize module on the target host in order to
copy, which uses rsync behind the scenes (hence the trailing slash
in the source argument).