roles/common: rework scripts to use s-nail for mail
Now all hosts can use s-nail, and we can use a mailrc file to set
the options. I've tested on Ubuntu 22.04 and CentOS Stream 8 with
unauthenticated SMTP and it works both interactively and via cron.
Cron was previously an issue on CentOS due to SELinux, but it works
out of the box now.