roles/dspace: rework dspace-angular.service
It seems my difficulty with start / stop was not misplaced. PM2
wants to save the state and this is problematic when managing the
starting and stopping via systemd. Running without the PM2 daemon
seems to work better.

Also, I am removing Type=forking, which will cause this to use the
simple type. Not sure about that, but it seems to be what people
are using with PM2 and systemd.

See: https://github.com/Unitech/pm2/issues/2914