roles/common: use s-nail as mail on Ubuntu
This is needed because unattended-upgrades has hardcoded support
for `mail`.

Note that there are a few problematic things for mail and this is
only one of them. The unattended-upgrades script also tries to use
sendmail (postfix) if it exists, which sends mail as "root@hostname"
and is probably filtered as spam or blocked outright depending on
the network, and even to send mail via s-nail we need to be white
listed by CGNET for SMTP. We basically just trust that security
updates will be installed by unattended-upgrades, and that we will
get *some* mails on failure from *some* hosts depending on above.

See: https://unix.stackexchange.com/questions/699230/how-to-use-s-nail-with-unattended-upgrades