roles/compute: revert to stock ssh_config
Use the stock CentOS 8 ssh_config and put our custom overrides in
the ssh_config.d directory that is included in the stock config.

According to the ssh_config man page, the *first* obtained value
is used, so our override file is called 01-ilri.conf to make sure
it is loaded before the 05-redhat.conf one.