roles/dspace: add robot user agent matching to nginx
Based on the COUNTER-Robots pattern matching with our overrides that
aren't in the upstream list yet. Note that nginx will use the first
matching pattern, so known good bot patterns must come first in the
config file. Also not that COUNTER-Robots patterns are expected to
be matched using case-insensitve regex, so I'm using "~*" here.