group_vars/dspace: adjust PostgreSQL effective_cache_size
The effective cache size is a hint to the PostgreSQL query planner,
not an allocation. Our database is only 2GB so the previous setting
of 37.5%, which amounted to ~11GB on our DSpace hosts, seems a bit
excessive (especially considering that the host runs other applica-
tions as well).

PostgreSQL's default is 4GB here and that might even be enough, but
I checked PGTune assuming we could allow PostgreSQL to use 8GB or so
in total and it recommended 6GB. Perhaps we'll end up going back to
the default after some monitoring and profiling.

See: https://wiki.postgresql.org/wiki/Tuning_Your_PostgreSQL_Server#effective_cache_size
See: https://pgtune.leopard.in.ua