roles/metrics: use restart instead of reload
The idea for this was that we don't need to restart the services
unless the service or configuration files are being created for
the first time. In the case that they were changed we only need to
reload them. It doesn't seem to work because the "diff" content in
the registered variable is always empty.

For now, in the interest of time, I will use the less cool restart.