roles/common: restart nftables in firewall update script
On Ubuntu and CentOS the nftables reload action is nearly identical
to the restart, but the signal is different in systemd. Now that we
are binding fail2ban to nftables with `PartOf`, we need that signal
to be a stop or start so they can come down and up together.