roles: rework nginx resolver
Set the resolver in the global http context because we need it for
resolving upstreams. In case the system's DNS is down, for example
apt decides to update libc and causes nginx and systemd-resolved to
restart at the same time, nginx will fail when it tries to resolve
any named upstreams.

We previously only used the resolver for OCSP stapling.