roles/dspace: revert nginx Angular cache
Users noticed a few side effects. For example, often users submit
an item and add a bitstream afterward. If the page has already been
cached it causes issues on the item page (bitstream shows up with a
lock icon). I think I will need to write an event consume to update
the cache when an item changes.