roles/common: remove CentOS appstream ignore for proj
Last month this broke updates due to epel's version being older, but
we stuck with it because proj-devel wasn't available in epel. Now
it seems proj-devel is in CentOS's crb stream so we should use that.

Note that this will break any R libraries that were built with it,
and we will need to re-install gdal.