roles/dspace: rework AReS proxy pass
Use the IP instead of an upstream, but specify the host header so
we get the right server on the other end. We do this now because
several times in the past few months the system's unattended upgr-
ades have caused network to be unavailable right when nginx is re-
starting and cause "host not found in upstream". Using the IP avo-
ids this and is not too risky because it hasn't changed in years.