roles/common: improve sshguard service
Add an override for the sshguard service on CentOS Stream like we
have for fail2ban elsewhere. First, we say that sshguard is PartOf
the nftables service so that restarts and stop events to nftables
are also propagated to sshguard. Second, harden the sshguard unit
to improve the security posture of the service a bit.